refactor: 도메인 위치 재구성
