+++ src/main/java/chess/domain/gameRoom/GameStatus.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.gameRoom;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.classification.Result;
