+++ src/main/java/chess/domain/gameRoom/dto/ChessGameRoomInfoDTO.java
@@ -1,4 +1,4 @@
-package chess.dto;
+package chess.domain.gameRoom.dto;
 
 public final class ChessGameRoomInfoDTO {
